Create input and checks if a given number is prime
In this code, the is_prime function checks whether a given number is prime or not. It starts by checking if the number is less than or equal to 1 because prime numbers are defined as positive integers greater than 1. Then, it iterates from 2 to the square root of the number and checks if there is any divisor for the given number. If a divisor is found, the function returns False, indicating that the number is not prime. Otherwise, it returns True.